verify backend, migrate to Render/Neon

+++ migrations/versions/3a70526dcb7c_add_version_to_notes.py
@@ -8,8 +8,8 @@ Create Date: 2026-05-19 16:42:33.728550
 from alembic import op
 import sqlalchemy as sa
 
+from util import column_exists
 
-# revision identifiers, used by Alembic.
 revision = '3a70526dcb7c'
 down_revision = 'edc83421847f'
 branch_labels = None
@@ -17,8 +17,13 @@ depends_on = None
 
 
 def upgrade():
-    op.add_column('notes', sa.Column('version', sa.Integer(), nullable=False, server_default='1'))
+    if not column_exists('notes', 'version'):
+        op.add_column(
+            'notes',
+            sa.Column('version', sa.Integer(), nullable=False, server_default='1'),
+        )
 
 
 def downgrade():
-    op.drop_column('notes', 'version')
+    if column_exists('notes', 'version'):
+        op.drop_column('notes', 'version')
